refactor: rename package to @angular-component/router
