+++ apps/example-app/src/app/books/containers/find-book-page.component.spec.ts
@@ -1,7 +1,7 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { ReactiveFormsModule } from '@angular/forms';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
-import { RoutingModule } from 'angular-routing';
+import { ComponentRouterModule } from '@angular-component/router';
 
 import { MockStore, provideMockStore } from '@ngrx/store/testing';
 
@@ -27,7 +27,7 @@ describe('Find Book Page', () => {
     TestBed.configureTestingModule({
       imports: [
         NoopAnimationsModule,
-        RoutingModule,
+        ComponentRouterModule,
         MaterialModule,
         ReactiveFormsModule,
       ],
