+++ apps/example-app/src/app/books/containers/view-book-page.component.spec.ts
@@ -1,5 +1,5 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { RouteParams } from 'angular-routing';
+import { RouteParams } from '@angular-component/router';
 
 import { provideMockStore, MockStore } from '@ngrx/store/testing';
 import { BehaviorSubject } from 'rxjs';
