+++ apps/example-app/src/app/core/containers/not-found-page.component.ts
@@ -1,5 +1,5 @@
 import { Component, ChangeDetectionStrategy } from '@angular/core';
-import { Router } from 'angular-routing';
+import { Router } from '@angular-component/router';
 
 @Component({
   selector: 'bc-not-found-page',
