+++ libs/router/ng-package.json
@@ -1,8 +1,8 @@
 {
   "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
-  "dest": "../../dist/libs/angular-routing",
+  "dest": "../../dist/libs/router",
   "assets": ["**/*.md", "./schematics/**/*.json"],
-  "whitelistedNonPeerDependencies": ["path-to-regexp", "query-string"],
+  "whitelistedNonPeerDependencies": ["query-string"],
   "lib": {
     "entryFile": "src/index.ts"
   }
