+++ libs/router/schematics/collection.json
