+++ libs/router/schematics/utils/config.ts
