+++ libs/router/src/lib/route-component.directive.ts
